feat: :lipstick: renaming components
